Correction warnings

+++ src/heapServer/clients.h
@@ -8,7 +8,7 @@ struct clientChain {
     struct clientChain *next;
 };
 
-extern int numClient;
+extern uint16_t numClient;
 extern int clientsConnected;
 extern struct clientChain *clients;
 extern struct replicationData *rep;
