+++ src/heapServer/servers.h
@@ -3,14 +3,14 @@
 
 struct serverChain {
 	int backup; // 0: main server, 1: backup server
-	uint16_t serverId;
+	uint8_t serverId;
     pthread_t threadId;
     int sock;
     char *serverAddress;
     struct serverChain *next;
 };
 
-extern int numServer;
+extern uint8_t numServer;
 extern int serversConnected;
 extern struct serverChain *servers;
 extern struct replicationData *rep;
